+++ launch/CHANGELOG.rst
@@ -2,6 +2,16 @@
 Changelog for package launch
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+3.1.0 (2023-09-08)
+------------------
+* Add maximum times for a process to respawn (`#696 <https://github.com/ros2/launch/issues/696>`_)
+* Contributors: Santti4go
+
+3.0.1 (2023-09-07)
+------------------
+* Add in a timeout for launch pytests. (`#725 <https://github.com/ros2/launch/issues/725>`_)
+* Contributors: Chris Lalancette
+
 3.0.0 (2023-08-21)
 ------------------
 * Fix remaining occurrences of "There is no current event loop" (`#723 <https://github.com/ros2/launch/issues/723>`_)